Committed to rolling our own query string parser
